bus: cleanup bus.go
